get some tests compiling
