+++ tokio/src/net/driver/mod.rs
@@ -124,13 +124,9 @@
 //! [`PollEvented`]: struct.PollEvented.html
 //! [`std::io::Read`]: https://doc.rust-lang.org/std/io/trait.Read.html
 //! [`std::io::Write`]: https://doc.rust-lang.org/std/io/trait.Write.html
-#[cfg(all(loom, test))]
-macro_rules! loom_thread_local {
-    ($($tts:tt)+) => { loom::thread_local!{ $($tts)+ } }
-}
 
 pub(crate) mod platform;
-mod reactor;
+pub(crate) mod reactor; // pub(crate) for slab tests
 mod registration;
 
 pub use self::reactor::{set_default, DefaultGuard, Handle, Reactor};
