+++ tokio/src/sync/task/atomic_waker.rs
@@ -1,3 +1,5 @@
+#![cfg_attr(any(loom, not(feature = "sync")), allow(dead_code, unreachable_pub))]
+
 use crate::loom::cell::CausalCell;
 use crate::loom::sync::atomic::{self, AtomicUsize};
