--- tokio/src/util/slab/pack.rs
@@ -1,87 +1,0 @@
-use crate::util::bit;
-
-pub(super) const WIDTH: usize = std::mem::size_of::<usize>() * 8;
-
-/// Trait encapsulating the calculations required for bit-packing slab indices.
-///
-/// This allows us to avoid manually repeating some calculations when packing
-/// and unpacking indices.
-pub(crate) trait Pack: Sized {
-    // ====== provided by each implementation =================================
-
-    /// The number of bits occupied by this type when packed into a usize.
-    ///
-    /// This must be provided to determine the number of bits into which to pack
-    /// the type.
-    const LEN: usize;
-    /// The type packed on the less significant side of this type.
-    ///
-    /// If this type is packed into the least significant bit of a usize, this
-    /// should be `()`, which occupies no bytes.
-    ///
-    /// This is used to calculate the shift amount for packing this value.
-    type Prev: Pack;
-
-    // ====== calculated automatically ========================================
-
-    /// A number consisting of `Self::LEN` 1 bits, starting at the least
-    /// significant bit.
-    ///
-    /// This is the higest value this type can represent. This number is shifted
-    /// left by `Self::SHIFT` bits to calculate this type's `MASK`.
-    ///
-    /// This is computed automatically based on `Self::LEN`.
-    const BITS: usize = bit::mask_for(Self::LEN);
-
-    /// The number of bits to shift a number to pack it into a usize with other
-    /// values.
-    ///
-    /// This is caculated automatically based on the `LEN` and `SHIFT` constants
-    /// of the previous value.
-    const SHIFT: usize = Self::Prev::SHIFT + Self::Prev::LEN;
-
-    /// The mask to extract only this type from a packed `usize`.
-    ///
-    /// This is calculated by shifting `Self::BITS` left by `Self::SHIFT`.
-    const MASK: usize = Self::BITS << Self::SHIFT;
-
-    fn as_usize(&self) -> usize;
-    fn from_usize(val: usize) -> Self;
-
-    fn pack(&self, to: usize) -> usize {
-        let value = self.as_usize();
-        debug_assert!(value <= Self::BITS);
-
-        (to & !Self::MASK) | (value << Self::SHIFT)
-    }
-
-    fn from_packed(from: usize) -> Self {
-        let value = bit::unpack(from, Self::MASK, Self::SHIFT as u32);
-        debug_assert!(value <= Self::BITS);
-        Self::from_usize(value)
-    }
-}
-
-impl Pack for () {
-    const BITS: usize = 0;
-    const LEN: usize = 0;
-    const SHIFT: usize = 0;
-    const MASK: usize = 0;
-
-    type Prev = ();
-
-    fn as_usize(&self) -> usize {
-        unreachable!()
-    }
-    fn from_usize(_val: usize) -> Self {
-        unreachable!()
-    }
-
-    fn pack(&self, _to: usize) -> usize {
-        unreachable!()
-    }
-
-    fn from_packed(_from: usize) -> Self {
-        unreachable!()
-    }
-}
